+++ devenv/tests/vrfv2plus/migration_test.go
@@ -2,7 +2,6 @@ package vrfv2plus
 
 import (
 	"context"
-	"fmt"
 	"math/big"
 	"strconv"
 	"testing"
@@ -33,8 +32,8 @@ const (
 
 func TestVRFv2PlusMigration(t *testing.T) {
 	t.Cleanup(func() {
-		_, cErr := framework.SaveContainerLogs(fmt.Sprintf("%s-%s", framework.DefaultCTFLogsDir, t.Name()))
-		require.NoError(t, cErr)
+		cleanupErr := products.CleanupContainerLogs(products.DefaultSettings())
+		require.NoError(t, cleanupErr, "failed to process cleanup container logs")
 	})
 
 	outputFile := "../../env-vrf2plus-out.toml"